pyproject.toml
[tool.poetry]
name = "gravity-common"
version = "1.0.0"
description = "Common library for Gravity Microservices - Reusable components"
authors = ["Gravity Elite Team <team@gravity.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
redis = "^5.0.1"
httpx = "^0.25.2"
prometheus-client = "^0.19.0"
python-json-logger = "^2.0.7"
orjson = "^3.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
black = "^23.12.1"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
